Commiting befor changing the structure yet again :)
The plan is to move away from an AST listing all different nodes, and
instead have a generic tree type we can move around in. This will need
to store what type the current node is and what the type of its children
is of course.

Hopefully this will work so that I can start concentrating on other features
like actually binding things, and maybe even typecheck. Another cool thing
would be to be able to eval and replace current focus with the result of one-
step evaluation.