Have an Expr tree and can Eval it.
It currently performs a very simple version of evaluation i.e no
  * alpha renaming
  * lookup vars in the env

It is a little cumbersome to work with the tree, not sure what to do
maybe using a better view type? But most of the changes I want to do
really should be done by calling functions. I'm thinking of adding
a type checker and then it will be important that trees are deleted
in the proper way.