Some changes to the ADT, revamped the input mechanism.
So most of the tree is now editable, which is kind of nice :) I
have also started to focus a little on how to actually edit. Now
it is much more smooth, altough it could of course be better.
So this commit before I merge back, a new branch will be created that
will focus on making better code of input handling.

Also maybe I should focus my branches better on specific changes but
hello I am newbie :)